+++ src/commands/emulators-exec.ts
@@ -14,6 +14,7 @@ export const command = new Command("emulators:exec <script>")
   .option(commandUtils.FLAG_EXPORT_ON_EXIT, commandUtils.DESC_EXPORT_ON_EXIT)
   .option(commandUtils.FLAG_VERBOSITY, commandUtils.DESC_VERBOSITY)
   .option(commandUtils.FLAG_UI, commandUtils.DESC_UI)
+  .option(commandUtils.FLAG_HOST, commandUtils.DESC_HOST)
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   .action((script: string, options: any) => {
     return Promise.race([shutdownWhenKilled(options), emulatorExec(script, options)]);