+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 - Fixed an issue in `apps:create` where App Store ID was always prompted for even when unnecessary.
 - Add `functions:lifecycle:list` and `functions:lifecycle:run` commands to view and run
   lifecycle hooks in isolation.
+- Add `--host-overrides` flag to dynamically override host and port for `emulators:start`
+  and `emulators:exec`.